requirements.txt
httpx>=0.27
python-dotenv>=1.0
openai>=1.40
matplotlib>=3.8
numpy>=1.26
datasets>=2.18
